+++ tests.go
@@ -17,7 +17,7 @@ type GenericTest struct {
 	ModifiedDate       *string             `json:"modifiedDate,omitempty"`
 	SavedEvent         *int                `json:"savedEvent,omitempty"`
 	SharedWithAccounts []SharedWithAccount `json:"sharedWithAccounts,omitempty"`
-	TestID             *int                `json:"testId,omitempty"`
+	TestID             *int64              `json:"testId,omitempty"`
 	TestName           *string             `json:"testName,omitempty"`
 	Type               *string             `json:"type,omitempty"`
 	// Fields unique to this test
